% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-linter.R
\name{build_markdown_scaffold}
\alias{build_markdown_scaffold}
\title{Write R script that functionalizes the knitr templates for a project type}
\usage{
build_markdown_scaffold(dev_dir = getwd(), project_type = "scRNA")
}
\arguments{
\item{dev_dir}{\code{character} naming path to write the R scripts}

\item{project_type}{\code{character} project type to write}
}
\value{

}
\description{
The functions will be appended to a file in \verb{<dev_dir>/scaffold_<project_type>.R}.
The roxygen skeleton should be modified as necessary to keep in date with the arguments.
}
