% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{call_intercalate}
\alias{call_intercalate}
\alias{call_intercalate_left}
\alias{call_intercalate_right}
\alias{meld_list_left}
\title{Splice in default arguments into a function}
\usage{
call_intercalate(f, ..., extra)

call_intercalate_left(f, ..., extra)

call_intercalate_right(f, ..., extra)

meld_list_left(x, y)
}
\arguments{
\item{f}{\code{function}}

\item{...}{key-value (named) arguments}

\item{extra}{named \code{list} of default arguments}

\item{x}{list}

\item{y}{list}
}
\value{
value of \code{f} with supplied arguments
}
\description{
Arguments in ... supercede in case of collisions with \code{extra}
}
\section{Functions}{
\itemize{
\item \code{call_intercalate_left}: don't warn with collision

\item \code{call_intercalate_right}: arguments in \code{extra} take presidence

\item \code{meld_list_left}: combine lists, preferentially taking elements from x if there are duplicate names
}}

\examples{
call_intercalate(sum, 3, 4, NA, extra = list(na.rm = TRUE))
call_intercalate_left(sum, 3, NA, na.rm = FALSE, extra = list(na.rm = TRUE))
call_intercalate_right(sum, 3, NA, na.rm = FALSE, extra = list(na.rm = TRUE))
meld_list_left(list(A=1, B=2), list(A = 0))
}
