% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-linter.R
\name{create_local_project}
\alias{create_local_project}
\alias{create_exampleproject}
\alias{create_temp_prefix}
\title{Create a local project for testing purposes}
\usage{
create_local_project(
  dir = tempdir(),
  env = parent.frame(),
  skeleton_args = list()
)

create_exampleproject(
  dir = tempdir(),
  env = parent.frame(),
  skeleton_args = list(project_type = "scRNA")
)

create_temp_prefix(prefix = tempfile(), env = parent.frame())
}
\arguments{
\item{dir}{\code{character} naming directory for local project}

\item{env}{\code{environment} to clean up after}

\item{skeleton_args}{\code{list} passed to \code{genesee_skeleton}}

\item{prefix}{prefix of files to be destroyed on exit}
}
\value{
\code{character} pointing to root of created project
}
\description{
Deleted after the environment named in \code{env} goes out of scope.
}
\section{Functions}{
\itemize{
\item \code{create_exampleproject}: create a project and copy in some additional files (eg example data)

\item \code{create_temp_prefix}: create and destroy files with a prefix
}}

