% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{file_mtime_exists}
\alias{file_mtime_exists}
\title{Return mtime if file exists}
\usage{
file_mtime_exists(file)
}
\arguments{
\item{file}{length one \code{character}}
}
\description{
Return mtime if file exists
}
\examples{
file_mtime_exists("Doesn't exist")
file_mtime_exists(system.file("DESCRIPTION", package = 'Genesee'))
}
