% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-linter.R
\name{generate_knitr_fun}
\alias{generate_knitr_fun}
\title{"Funcionalize" parameters in a knitr script}
\usage{
generate_knitr_fun(name, pars)
}
\arguments{
\item{name}{\code{character} name of function (name of script)}

\item{pars}{\code{list} of knitr parameters as returned by \code{knitr::knit_params}}
}
\value{
\code{function} with formals set according to \code{pars} and a body with a usage message.
}
\description{
"Funcionalize" parameters in a knitr script
}
