% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-setup.R
\name{genesee_skeleton}
\alias{genesee_skeleton}
\title{Initialize a Genesee project}
\usage{
genesee_skeleton(
  genesee_root,
  investigator,
  project_title,
  init_templates = TRUE,
  use_renv = FALSE,
  project_type = "scRNA",
  authors = "default",
  force = FALSE,
  navigate_rawdata = TRUE
)
}
\arguments{
\item{genesee_root}{\code{character} naming a containing all genesee projects}

\item{investigator}{\code{character} naming PI of project}

\item{project_title}{\code{character} naming project}

\item{init_templates}{\code{logical} Should template markdown, shell scripts, etc, be copied?}

\item{use_renv}{\code{logical} Should renv be initialized?}

\item{project_type}{\code{character}, one of \code{scRNA}, \code{RNA} or \code{scVDJ}}

\item{authors}{\code{character} containing the authors/editors of the project}

\item{force}{\code{logical} force creation of a project}

\item{navigate_rawdata}{\code{logical} should we shift focus to the rstudio terminal opened on the rawdata directory}
}
\value{
TRUE on success
}
\description{
This creates a set of directories, activates \code{renv}, copies down templated
shell scripts.
The next step is to run make_sample_sheet.
}
