% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-setup.R
\name{interpolate_and_copy}
\alias{interpolate_and_copy}
\title{Interpolate and copy template files}
\usage{
interpolate_and_copy(
  project_directory = NULL,
  init_markdown = TRUE,
  init_sh = TRUE,
  init_gitfiles = TRUE,
  init_R = TRUE,
  type = NULL
)
}
\arguments{
\item{project_directory}{directory of the Genesee project. If set to \code{NULL} then will use working directory as long as a `config.yml`` is present.}

\item{init_markdown}{\code{logical} Should markdown be copied?}

\item{init_sh}{\code{logical} Should shell script be copied?}

\item{init_gitfiles}{\code{logical} Should files managing git configuration be copied (eg .gitignore)}

\item{init_R}{\code{logical} should R scripts be copied}

\item{type}{\code{character} The type of project.  One of \code{scRNA} and \code{scVDJ}.  If set to \code{NULL} then will attempt to read from \code{config.yml}.}
}
\value{
\code{logical} upon success
}
\description{
Interpolate and copy template files
}
