% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-setup.R
\name{make_sample_sheet}
\alias{make_sample_sheet}
\title{Make a sample sheet reflecting the samples apparent in rawdata}
\usage{
make_sample_sheet(
  project_directory = NULL,
  write = TRUE,
  overwrite = FALSE,
  type = NULL
)
}
\arguments{
\item{project_directory}{directory of the Genesee project. If set to \code{NULL} then will use working directory as long as a `config.yml`` is present.}

\item{write}{\code{logical} should the sample sheet be written to file, or just returned for further messaging}

\item{overwrite}{\code{logical} should an existing sheet be overwritten?}

\item{type}{\code{character} The type of project.  One of \code{scRNA} and \code{scVDJ}.  If set to \code{NULL} then will attempt to read from \code{config.yml}.}
}
\value{
path of sample sheet (for additional editing) or sample sheet itself
}
\description{
The sample sheet \code{all_samples_files.csv} maps between sequencing runs and
other covariates.
}
\details{
At the moment, Cellranger mRNA (\code{type = 'scRNA'}) and Cellranger VDJ (\code{type = 'scVDJ'}) are supported.
}
