% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-matrix-methods.R
\name{split_matrix_col}
\alias{split_matrix_col}
\alias{lapply_attr}
\alias{split_matrix_row}
\alias{unsplit_matrix}
\title{Split a matrix in a list of matrices by levels of a factor / integer}
\usage{
split_matrix_col(m, i)

lapply_attr(X, FUN, ...)

split_matrix_row(m, i)

unsplit_matrix(x)
}
\arguments{
\item{m}{\code{matrix}-like}

\item{i}{\code{factor}-like}

\item{X}{a vector (atomic or list) or an \code{\link[base]{expression}}
    object.  Other objects (including classed objects) will be coerced
    by \code{base::\link[base]{as.list}}.}

\item{FUN}{the function to be applied to each element of \code{X}:
    see \sQuote{Details}.  In the case of functions like
    \code{+}, \code{\%*\%}, the function name must be backquoted or quoted.}

\item{...}{optional arguments to \code{FUN}.}

\item{x}{a list, typically returned from \code{lapply()}.}
}
\value{
list of matrices of \code{length(unique(i))}
}
\description{
This function also works with \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment()}} / \code{\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment::SingleCellExperiment()}}
}
\section{Functions}{
\itemize{
\item \code{lapply_attr}: lapply and preserve attributes.  Necessary for \code{unsplit_matrix}.

\item \code{split_matrix_row}: split a matrix by row

\item \code{unsplit_matrix}: reverse a split (cbind/rbind) and reordering
}}

\examples{
m = matrix(1:12, nrow = 3)
s_col = split_matrix_col(m, c('A', 'A', 'B', 'B'))
s_row = split_matrix_row(m, c('B', 'A', 'B'))
scaled = lapply_attr(s_col, function(x) x/sum(x))
unsplit_matrix(scaled)
stopifnot(all.equal(unsplit_matrix(s_row), m))
}
